+++ windows/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   firebase_core
+  firebase_remote_config
   geolocator_windows
   url_launcher_windows
 )